Move treenode and sidemenu out of models folder
